Split Thread.cs into ThreadFactory.cs and Thread_1.cs
